Commands: Added private command management
For other stuff, like UIs, the engine would have a private
structure with variables that (probably) shouldn't be editable by the client.
Commands used to not do this, but with this change, it now does.

This fixes many memory leaks, and the engine no longer
complains about a memory leak.